secure-data-processing-service/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secure-data-processing-service"
version = "0.1.0"
description = "Security-first data processing service with encryption, MFA, auditing, and compliance reporting"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.7.0",
  "PyJWT>=2.8.0",
  "cryptography>=42.0.0",
  "httpx>=0.27.0",
  "python-multipart>=0.0.9",
  "PyYAML>=6.0.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools.packages.find]
include = ["app*"]
